add in logic to shut down the istio-proxy container
When istio is enabled, the gitlab-toolbox-backup job runs.  However, the
job doesn't not complete.

This fix adds in logic to take advantage of the extraArgs variable in
the upsteam backup job chart and clean up when istio and the backup cron
are both enabled.  Other job charts in the gitlab package did not expose
an extraArgs variable, so those charts had to be modified.

Existing extraArgs are passed through if present.